% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse.page.R
\name{analyse.page}
\alias{analyse.page}
\title{fonction affichant une page d'analyse}
\usage{
analyse.page(File = File, ech = seq(1, 2), Dose = c(0, 30, 50, 70, 0, 30),
  ph0 = seq(1, 4), NomEch = c(ech1 = "ech1", ech2 = "ech2", ech3 = "ech3",
  ech4 = "ech4"), Unique = FALSE)
}
\arguments{
\item{File}{fichier Bin}

\item{ech}{numero echantillon}

\item{Dose}{doses en secondes}

\item{ph0}{prechauffe}

\item{NomEch}{nom de l echantillon}

\item{Unique}{TRUE si la dose equivalente est independant des prechauffe}
}
\description{
fonction affichant une page d'analyse
}
