% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OSLpack-package.R
\docType{data}
\name{Anatolian1}
\alias{Anatolian1}
\title{OSL data of Anatolian figurine (genuine)}
\format{
A Risoe.BINFileData object
}
\source{
1203214.binx OSL SAR SH112481, SH112482 {N; 200sb; 250sb; 300sb; 0sb; 200sb}; 1503211.binx OSL SAR SH112481{100sb}
}
\usage{
data(Anatolian1)
}
\description{
An example of actual data from a OSL measurement
}
\details{
This data set gives the OSL glow curve of SH112481
}
\keyword{datasets}
