% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OSLpack-package.R
\docType{data}
\name{Anatolian2}
\alias{Anatolian2}
\title{OSL data of Anatolian figurine (fake)}
\format{
A Risoe.BINFileData object
}
\source{
1203214.binx TL OSL SAR SH112481, SH112482 {N; 200sb; 250sb; 300sb; 0sb; 200sb}
}
\usage{
data(Anatolian2)
}
\description{
An example of actual data from a OSL measurement
}
\details{
This data set gives the OSL glow curve of SH112482
}
\keyword{datasets}
