% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesCal.R
\name{BayesCal}
\alias{BayesCal}
\title{BayesCal}
\usage{
BayesCal(Sn, ph, debug)
}
\arguments{
\item{Sn}{[list] (**required**) efficiencies S table (see Sn function)}

\item{ph}{[numeric] (**required**): selected preheat}

\item{debug}{[logical] (**required**) TRUE debug the WinBug code}
}
\value{
WinBUGS simulation results (see R2WinBUGS::bugs help page)
}
\description{
De calculation using gibbs sampler (WinBUGS code call)
}
