% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GrowthCurve.plot.R
\name{GrowthCurve.plot}
\alias{GrowthCurve.plot}
\title{GrowthCurve.plot}
\usage{
GrowthCurve.plot(Sn, ph0 = seq(1, 4), Unique = FALSE, debug)
}
\arguments{
\item{Sn}{[list] efficiencies S table (see Sn function)}

\item{ph0}{[numeric] (**with default**): selected preheat}

\item{Unique}{[logical]  (**with default**)  TRUE a single De, independent from preheat temperature}

\item{debug}{[logical] (**required**) TRUE debug the WinBug code}
}
\value{
a plot

WinBUg results object
}
\description{
growthcurve plot at a single preheat temperature
}
