% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiOSL.plot.R
\name{MultiOSL.plot}
\alias{MultiOSL.plot}
\title{MultiOSL.plot}
\usage{
MultiOSL.plot(file = file, ech = 1, OSL = 2, Dose = c(0, 30, 50, 70, 0,
  30), nom = nomFile, ph0 = seq(1, 4), NomEch = c(ech1 = "ech1", ech2 =
  "ech2", ech3 = "ech3", ech4 = "ech4"))
}
\arguments{
\item{NomEch}{}
}
\description{
MultiOSL.plot
}
