% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiplot.OSL.R
\name{multiplot.OSL}
\alias{multiplot.OSL}
\title{multiplot.OSL}
\usage{
multiplot.OSL(
  file,
  ech = 1,
  OSL = 2,
  Dose = c(0, 30, 50, 70, 0, 30),
  nomFile = "nomFile",
  ph0 = seq(1, 4),
  NumInv = c(ech1 = "ech1", ech2 = "ech2", ech3 = "ech3", ech4 = "ech4")
)
}
\arguments{
\item{file}{[Risoe.BINfileData-class],[list] (**required**) the BIN/BINX file(s)}

\item{ech}{[numeric] (**with default**) the sample number}

\item{OSL}{[numeric] (**with default**): 1: IR-OSL; 2: BL-OSL}

\item{Dose}{[numeric], [list] (**with default**): doses in seconds}

\item{nomFile}{[character] (**with default**) name of the BIN/BINX file}

\item{ph0}{[numeric] (**with default**): selected preheat}

\item{NumInv}{[character] (**with default**) the name of the sample (eg. inventory number)}
}
\value{
a page with multiplot
}
\description{
OSL glowcurve for various preheat
}
\examples{
data(Anatolian2, envir = environment())
file<-Anatolian2$FILE
multiplot.OSL(file,ech=2,OSL=1,Dose=c(0,200,250,300,0,200),ph0=seq(1,4),NumInv=c("SH112482"))

}
