% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OSLpack-package.R
\docType{package}
\name{OSLpack-package}
\alias{OSLpack}
\alias{OSLpack-package}
\title{OSLpack: Analyze OSL using Bayesian}
\description{
Various functions to analyze and plot OSL measurements using Bayesian approach (BUGS code)
}
\author{
\strong{Maintainer}: Antoine Zink \email{antoine.zink@culture.gouv.fr} (https://orcid.org/0000-0001-7146-1101)

}
\keyword{internal}
