% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadFile.R
\name{ReadFile}
\alias{ReadFile}
\title{ReadFile
load one or several BIN/BINX files}
\usage{
ReadFile(list = "liste fichier.txt", n = 1, embedData = TRUE)
}
\arguments{
\item{list}{[string] (**with default**) name of files list (default "liste fichier.txt")}

\item{n}{[numeric] (**with default**) number of files (default n=1)}

\item{embedData}{[logical]  (**with default**)  TRUE (default) retrieves data, otherwise (FALSE) returns only file names}
}
\value{
depending on the value of *embedData*

TRUE (default value)

FILE object of class "list" containing S4 Risoe.BINFileData objects

NFILE File name (*.bin/*.binx)

FALSE

File name (*.bin/*.binx)
}
\description{
ReadFile
load one or several BIN/BINX files
}
\examples{
\dontrun{
File<-ReadFile(n=1)
file<-File$FILE
nomFile<-File$NFILE
}
}
