% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sn.R
\name{Sn}
\alias{Sn}
\title{Sn}
\usage{
Sn(file, ech, OSL, Dose = c(0, 50, 80, 110, 0, 50), TypLum = c("LIR", "LBL",
  "TIR", "TBL"), Temp = c("250째C", "275째C", "300째C", "325째C"),
  ph0 = seq(1, 4), signal.integral = seq(6, 10),
  background.integral = seq(85, 104))
}
\arguments{
\item{file}{[Risoe.BINfileData-class] or [list] (*required*): BIN/BINX file}

\item{ech}{[numeric] (*required*): sample number}

\item{OSL}{[numeric] (*required*): 1: IR-OSL; 2: BL-OSL}

\item{Dose}{[numeric] or [list] (*with default*): doses in seconds}

\item{TypLum}{[character] or [list] (*with default*): type of luminescence LIR; LBL; TIR; TBL}

\item{Temp}{[character] or [list] (*with default*): preheat temperature}

\item{ph0}{[numeric] (*with default*): selected preheat}

\item{signal.integral}{[vector] (*with default*): vector with the limits for the signal integral.}

\item{background.integral}{[vector] (*with default*): vector with the bounds for the background integral.}
}
\description{
determines the sensitivity values for OSL
}
