% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse.plot.R
\name{analyse.plot}
\alias{analyse.plot}
\title{analyse.plot}
\usage{
analyse.plot(
  file,
  nom = nomFile,
  ech = 1,
  OSL = 2,
  Dose = c(0, 30, 50, 70, 0, 30),
  Unique = FALSE,
  ph0 = seq(1, 4),
  NomEch = c(ech1 = "ech1", ech2 = "ech2", ech3 = "ech3", ech4 = "ech4"),
  signal.integral = seq(6, 10),
  background.integral = seq(85, 104),
  debug = FALSE,
  brate = 0.1,
  ...
)
}
\arguments{
\item{file}{[Risoe.BINfileData-class],[list] (**required**) the BIN/BINX file(s)}

\item{nom}{[character] (**optional**) BIN/BINX file name. Use nomFile as global variable}

\item{ech}{[numeric] (**with default**) the sample number}

\item{OSL}{[numeric] (**with default**): 1: IR-OSL; 2: BL-OSL}

\item{Dose}{[numeric], [list] (**with default**): doses in seconds}

\item{Unique}{[logical]  (**with default**)  TRUE a single De, independent from preheat temperature}

\item{ph0}{[numeric] (**with default**): selected preheat}

\item{NomEch}{[character] (**optional**) sample name}

\item{signal.integral}{[vector] (**with default**): vector with the limits for the signal integral.}

\item{background.integral}{[vector] (**with default**): vector with the bounds for the background integral.}

\item{debug}{if \code{FALSE} (default), \pkg{WinBUGS} is closed automatically
    when the script has finished running, otherwise \pkg{WinBUGS} remains open
    for further investigation}

\item{brate}{[numeric] (**with default**) dose rate of the beta source (default value 0.1 Gy/s)}

\item{...}{further arguments that will be passed to the function BayesCal}
}
\value{
several plots

WinBUg results objects
}
\description{
analyse OSL-SAR measurement at various preheat and plot the result
}
\examples{
data(Anatolian2, envir = environment())
file<-Anatolian2$FILE

\dontrun{
png("Analyse BL SH112482.png",width=1063,height=1500, units = "px", pointsize = 24)
analyse.plot(file,ech=1,OSL=2,Dose=c(0,200,250,300,0,200),ph0=c(1,2,3,4),
                  Unique=FALSE,NomEch=c("figurine Anatolie SH112482"))
dev.off()
}

}
