% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GrowthCurve.plot.R
\name{GrowthCurve.plot}
\alias{GrowthCurve.plot}
\title{GrowthCurve.plot}
\usage{
GrowthCurve.plot(Sn, ph0 = seq(1, 4), Unique = FALSE, debug)
}
\arguments{
\item{debug}{}
}
\description{
GrowthCurve.plot
}
