% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadFile.R
\name{ReadFile}
\alias{ReadFile}
\title{ReadFile
load one or several BIN/BINX files}
\usage{
ReadFile(list = "liste fichier.txt", n = 1)
}
\arguments{
\item{list}{name of files list (default "liste fichier.txt")}

\item{n}{number of files (default n=1)}
}
\value{
FILE object of class "list" containing S4 Risoe.BINFileData objects

NFILE File name (*.bin/*.binx)
}
\description{
ReadFile
load one or several BIN/BINX files
}
\examples{
\dontrun{
File<-ReadFile(n=1)
file<-File$FILE
nomFile<-File$NFILE
}
}
